refactor: Export mqlArray from helper.js
