+++ js/modules/helpers.js
@@ -27,3 +27,11 @@ export function throttle(func, ms){
 
 	return wrapper;
 }
+
+const mql320 = window.matchMedia("(min-width: 320px)");
+const mql480 = window.matchMedia("(min-width: 480px)");
+const mql768 = window.matchMedia("(min-width: 768px)");
+const mql1000 = window.matchMedia("(min-width: 1000px)");
+const mql1920 = window.matchMedia("(min-width: 1920px)");
+
+export const mqlArray = [mql320, mql480, mql768, mql1000, mql1920];
